requirements.txt
python-dotenv
requests
tqdm
pandas